Finish refactor of DataPoint
