+++ RNAFoldAssess/models/__init__.py
@@ -1,5 +1,5 @@
 from .data_point import *
-from .data_point2 import *
+from .data_point import *
 from .eterna_data_point import *
 from .data_point_decorator import *
 from .data_point_from_crystal import *
