+++ tests/test_rnafold.py
@@ -7,7 +7,7 @@ from RNAFoldAssess.models import DSCI
 class TestRNAFold:
     # Testing with C009C
     base_data_path = "./tests/fixtures"
-    datum = DataPoint.factory(f'{base_data_path}/C009C.json')[0]
+    datum = DataPoint.factory_from_json(f'{base_data_path}/C009C.json')[0]
     model = RNAFold()
     def test_prediction_with_fasta_file(self):
         input_file_path = self.datum.to_fasta_file()
@@ -20,7 +20,7 @@ class TestRNAFold:
 
     def test_prediction_with_string_input(self):
         base_data_path = "./tests/fixtures"
-        datum = DataPoint.factory(f'{base_data_path}/C009C.json')[0]
+        datum = DataPoint.factory_from_json(f'{base_data_path}/C009C.json')[0]
         sequence = datum.sequence
         self.model.execute(sequence)
         prediction = self.model.get_ss_prediction()
